# src/main/resources/static/v3/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "U Service API",
    "description": "A service for managing `U` (user) instances.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Localhost server"
    }
  ],
  "tags": [
    {
      "name": "u-controller"
    }
  ],
  "paths": {
    "/u": {
      "get": {
        "tags": [
          "u-controller"
        ],
        "summary": "Get all U instances",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/U"
                  }
                }
              }
            }
          },
          "404": {
            "description": "desc"
          }
        }
      },
      "post": {
        "tags": [
          "u-controller"
        ],
        "summary": "Create a new U instance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/U"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "U instance created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "f1251a28-50e8-4f74-878e-ac6db1b1591a",
                  "email": "example@example.com",
                  "nickname": "ExampleUser",
                  "dateOfBirth": "1990-01-01",
                  "nativeLanguage": "english",
                  "subscription": "free",
                  "registrationMethod": "email",
                  "isActive": true,
                  "verificationCode": "ABC123",
                  "words": [
                    "c2d5b7e8-b7b9-4fb4-8248-39904ccce3b7"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "U instance not created, invalid request body provided",
            "example": ""
          }
        }
      }
    },
    "/u/{id}": {
      "get": {
        "tags": [
          "u-controller"
        ],
        "summary": "Get a specific U instance by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "5fa51ebe-10f8-4b14-9409-996ef9140395"
          }
        ],
        "responses": {
          "200": {
            "description": "U instance retrieved successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": "5fa51ebe-10f8-4b14-9409-996ef9140395",
                  "email": "example@example.com",
                  "nickname": "ExampleUser",
                  "dateOfBirth": "1990-01-01",
                  "nativeLanguage": "English",
                  "subscription": "Free",
                  "registrationMethod": "Email",
                  "isActive": true,
                  "verificationCode": "ABC123",
                  "words": [
                    "c2d5b7e8-b7b9-4fb4-8248-39904ccce3b7"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "u-controller"
        ],
        "summary": "Partially update a specific U instance by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "f1251a28-50e8-4f74-878e-ac6db1b1591a"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "nickname": "UpdatedUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "U instance updated successfully.",
            "content": {
              "application/json": {
                "example": {
                  "id": "f1251a28-50e8-4f74-878e-ac6db1b1591a",
                  "email": "example@example.com",
                  "nickname": "UpdatedUser",
                  "dateOfBirth": "1990-01-01",
                  "nativeLanguage": "English",
                  "subscription": "Free",
                  "registrationMethod": "Email",
                  "isActive": true,
                  "verificationCode": "ABC123",
                  "words": [
                    "c2d5b7e8-b7b9-4fb4-8248-39904ccce3b7"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "u-controller"
        ],
        "summary": "Delete a specific U instance by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "5fa51ebe-10f8-4b14-9409-996ef9140395"
          }
        ],
        "responses": {
          "204": {
            "description": "U instance deleted successfully.",
            "content": {
              "application/json": {
                "example": {}
              }
            }
          },
          "500": {
            "description": "U instance not found.",
            "content": {
              "application/json": {
                "example": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "U": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "nativeLanguage": {
            "type": "string"
          },
          "subscription": {
            "type": "string"
          },
          "registrationMethod": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "verificationCode": {
            "type": "string"
          },
          "words": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      }
    }
  }
}